Create dijkstra's algo
we initialize the distance from source to other vertices as infinite. sp[] array is will tell us if the vertex has been visited or not. In function mindist we initialise. if the distance of the vertex is less than max size we return the index. In dijkstra function, we have an array of vertices and an array of visited vertices. All vertices are intially set to infinte and flase. dist[source] is 0. we enter a loop for dist[]. mindist function is called, which will return the index of the min distance. sp[u] is true because the vertex has now been visited. once again, we enter a loop for the same array. If the distance to one vertext from the source + the distance in the original graph is less than the actual, direct distance, that is the min distance.